Refactor CV utilities and config handling
